CommandHandler command switch cleared

+++ sources/communication/communicators/mavlink/handlers/command/command_handler.h
@@ -34,8 +34,8 @@ namespace comm
         void sendSetMode(quint8 mavId, domain::vehicle::Mode mode);
         void sendCurrentItem(quint8 mavId, quint16 seq);
         void sendNavTo(quint8 mavId, double latitude, double longitude, float altitude);
-        void sendChangeAltitude(quint8 mavId, double altitude);
-        void sendManualControl(int vehicleId, double pitch, double roll, double yaw, double thrust);
+        void sendChangeAltitude(quint8 mavId, float altitude);
+        void sendManualControl(int vehicleId, float pitch, float roll, float yaw, float thrust);
 
     private:
         class Impl;
